+++ CHANGELOG.markdown
@@ -4,3 +4,4 @@
  - Added new `Gravatar\Image::getImage()` return type: __Markdown__
  - Deleted all `Gravatar\Profile` constans, added getters
  - Added background getters
+ - `Gravatar\Profile::loadProfile()` now throws `RuntimeException` when failed to load